Add files via upload
With the group assignments; in Assignment 1, we were required to investigate Java Frameworks and Libraries - define Java Frameworks and Java Libraries, Compare and contrast between frameworks and libraries, pros and cons (of frameworks and libraries), and conclude with a summary. In Assignment 2, we focused on the Spring Framework; defining the Spring Framework, the functionality of the Spring Framework/how Spring Framework Works, Inversion of Control (IoC) and Dependency Injection (DI), and the Advantages of Spring Framework.